Initial Commit. Added the main.py file.
This program finds the minimum number of clicks that it takes to get from one arbitrary Wikipedia page to another.
Unfortunately, assuming that the number of links on any given Wikipedia page is somewhat the same, the problem grows
exponentially.

This version is broken and only features a function that will return 1 if the Wikipedia page provided directly contains
the target link. If the link is two clicks away, then the program will probably not find it.

-Zach