First Working Version. Essentially a brute force program.
This version will successfully find the number of clicks, given enough time.
This version is great for finding the minimum number of clicks if the topics are similar.
i.e. How many clicks from "Jelly Doughnut" to "Apricot Jam"?
Again, due to exponential growth, this program typically takes a long time to find the number of clicks
if it is over two.

This version implemented the "title_list" and the "title_text" functions.
The "clickfinder" function is no longer recursive.

-Zach